refactor: Refactor provider injection blocklist to avoid using regex (#23134)
Fixes two bugs that occurred as the result of using regex to
identify URLs in our content script blocklist.

The first issue is that of escaping the first `.` found in a
URL when using the inpage blocklist. This meant that entries such as
`ani.gamer.com.tw` would have their first period escaped for regex
parsing, but subsequent periods were treated as regex wildcards. This
could lead to and unintentionally matching on URLs such as
`ani.gamerxcom.tw` etc.

The second issue is that of missing a leading anchor `^` in the
regex expression. This means that we would block the domain if the
matched string occurred anywhere in the URL. For example,
`https://google.com?search=uscourts.gov` would be blocked since
it ended in `uscourts.gov`.

To avoid future regex complexities, this code has been refactored to use
built-in JavaScript URL parsing instead.

---------

Co-authored-by: legobeat <109787230+legobeat@users.noreply.github.com>