fix: crash when connecting to multiple Snaps at once (#23646)
## **Description**

Fixes a crash when a user tries to connect to multiple Snaps at once. As
part of a localization change we accidentally used the wrong selector to
get the snap name. This PR swaps that selector for one that more
gracefully handles the Snap not being present in the state.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/23646?quickstart=1)