build: remove `brfs` from build processes (#24000)
The `brfs` package was used to transform calls to `fs.readFile*` into
inlined files as strings during build/compilation. We no longer make use
of `fs.readFile*` in our code, so the dependency is no longer required.
While it doesn't cause issues to leave it, the package is slow as it
must parse the contents into an AST and then statically analyze the tree
looking for relevant calls to `fs.readFile*`. This overhead is unwanted,
so I've removed it from our build process.