ci: fix `Validate Conventional Commit Title` action (#24356)
The `Validate Conventional Commit Title` action was only running when
the PR was opened (and if the title or base are edited, or the PR was
closed then re-opened). It was _not_ when the branch had a new commit
added; in that case, the previous run's status would just be completely
ignored by GitHub.

Adding "synchronize" fixes it (as the installation instructions specify:
https://github.com/amannn/action-semantic-pull-request?tab=readme-ov-file#installation)

note: we probably don't need "reopened", but I don't know why that was
added in the first place so I'm not removing it.

I tested this here:
https://github.com/MetaMask/metamask-extension/pull/24355/commits

Check out the statuses for each commit (hover over the red X). The
action ran on the first commit, as expected, and it did not ran on the
next two, but the last commit added `synchronize` (just like in this
PR), and it ran again.

<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.


## **Description**


[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/24356?quickstart=1)

## **Related issues**

Fixes:

## **Manual testing steps**

1. Go to this page...
2.
3.

## **Screenshots/Recordings**

### **Before**

### **After**

## **Pre-merge author checklist**

- [ ] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.
-->